+++ lib/screens/customer/home/profile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../components/oilwale_theme.dart';
+import 'package:Oilwale/components/oilwale_theme.dart';
 
 class ProfileScreen extends StatelessWidget {
   @override
