+++ lib/screens/customer/index.dart
@@ -1,4 +1,4 @@
 export 'createaccount.dart';
 export 'home.dart';
-export 'login.dart';
-export 'splash.dart';
+export '../login.dart';
+export '../splash.dart';
