+++ lib/screens/garage/offerdetails.dart
